+++ contribs/falkhash/falkhash.h
@@ -4,7 +4,7 @@
 extern "C" {
 #endif
 
-uint64_t falkhash64(const void * key, int len, uint64_t seed);
+uint64_t falkhash64(const void * key, size_t len, uint64_t seed);
 
 #if defined (__cplusplus)
 }
